requirements.txt
cocos2d
six
pyglet